requirements.txt
beautifulsoup4==4.8.2
cffi==1.13.2
Django==2.2.9
django-ckeditor==5.8.0
django-crispy-forms==1.8.1
django-js-asset==1.2.2
docxtpl==0.6.3
html5lib==1.0.1
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.19
pylokit==0.8.1
PyPDF2==1.26.0
python-decouple==3.3
python-docx==0.8.7
pytz==2019.3
reportlab==3.5.34
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
webencodings==0.5.1
xhtml2pdf==0.2.3
